+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [v2.5.0](https://github.com/riodelphino/macime.nvim/compare/v2.4.1...v2.5.0) (2026-02-21)
 
+* **feat:** Adapt to `--log-level` option in `macime` v4.3.0
 * **feat:** Adapt to `--cjk-delay` option in `macime` v4.3.0
 
 ## [v2.4.1](https://github.com/riodelphino/macime.nvim/compare/v2.4.0...v2.4.1) (2026-02-17)